feat: ExteriorColorResponseDTO 코드 분리
