+++ iOS/Catalog/Catalog/ExternalColorSelection/Data/ExteriorColorRepository.swift
@@ -23,11 +23,3 @@ final class ExteriorColorRepository: ExteriorColorRepositoryProtocol {
 
 }
 
-struct ExteriorColorResponseDTO: Decodable {
-  let id: Int?
-  let name: String?
-  let choiceRatio: Int?
-  let price: Int?
-  let hexCode: String?
-  let images: [String]?
-}
